requirements.txt
amqp==2.6.0
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.0
certifi==2020.4.5.2
chardet==3.0.4
Django==3.0.14
idna==2.9
kombu==4.6.10
psycopg2-binary==2.8.6
pytz==2020.1
redis==3.4.1
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
djangorestframework
pytest-django
django-cors-headers==3.4.0